German translation a more flexible amount parser
Adding German translation. On the way, I realized that amounts will be
separated with the . (dot) sign. This is right for the US, but other
countries/languages use other symbols as a separator. The translation
file already handle this, but the code was not ready to handle amounts
with , (comma) or an EUR symbol at the end. For that, there is a new
helper method, which will transform the input string (an maybe localized
amount) to a valid float, so it can be used to do calculations.

On the way I've cleaned up some typos.