prepare internals for robinhood integration
working on robinhood integration on the backend. if someone can do the frontend that'd be dope. I was thinking of a section under RECOMMENDATIONS called INVESTMENTS and under that two labels (Portfolio and Buying Power) which I would fill in on the backend. Possibly also an interchangable login/logout button below.