Added fill property to all styles
The income and expense pop out charts were dependent on the body
styling. The body styling did not have a fill property and only used
the color property. As a result some of the SVG elements were using the
default fill properties which caused them to not be visible against dark
backgrounds. Adding the fill property in the body styling enabled the
SVG elements to inheriet the var(--forground-color).